EditPointTool,EditLineTool: Enlarge object frame
In edit mode, the object's surrounding frame can be used for moving the
object.  However, in current implementation the frame may overlap with
some of the object vertices. Then dragging the frame at the place where
it collides with a vertex results in unwanted vertex movement. This
commit makes the surrounding frame larger so that distinction between
the object move handle and the vertices is clear.

Closes GH-980